requirements.txt
fastapi
uvicorn
google-generativeai
gtts
pydantic
python-multipart
